Refactor: 리팩터링
