--- src/main/java/domain/Lotto.java
@@ -1,11 +1,0 @@
-package domain;
-
-import java.util.List;
-
-public record Lotto(List<Integer> numbers) {
-    public Lotto {
-        numbers = numbers.stream()
-                .sorted()
-                .toList();
-    }
-}
