+++ src/main/java/domain/strategy/ManualGenerate.java
@@ -1,6 +1,7 @@
 package domain.strategy;
 
-import domain.Lotto;
+import domain.model.Lotto;
+import domain.util.LottoUtils;
 
 import java.util.Collections;
 import java.util.List;
@@ -9,6 +10,6 @@ public class ManualGenerate implements LottoStrategy{
     @Override
     public Lotto generateLotto(List<Integer> manualNumbers){
         Collections.sort(manualNumbers);
-        return new Lotto(manualNumbers);
+        return new Lotto(LottoUtils.convertToLottoNumbers(manualNumbers));
     }
 }
